// frontend/src/generated/models/Transaction.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Transaction = {
    account_id: string;
    amount: number;
    category?: string | null;
    created_at: string;
    description: string;
    id: string;
    memo?: string | null;
    payee?: string | null;
    pending?: boolean | null;
    posted_date: string;
    simplefin_id?: string | null;
    transaction_date: string;
};